APM Version 1.3
Support for class sockets in basic.cfg
UI changes.
Fixed 32 bit int crash.
Automatic support for \ServerPath\!Workshop folder.
Local mods in !Workshop or ServerPath folder won't have their paths displayed.
Performance/Profiling Branch support.
Missions packed as .pbo's will be displayed as such.
Reduced maximum amount of Headless Clients supported to ACEX maximum.
UI changes from Keelah's version.
UserData saved in \AppData\Local\Artamidae.
Defaults tweaked to match APM's settings.

+++ FASTER/Models/SteamWebApi.cs
@@ -18,7 +18,7 @@ namespace FASTER.Models
         private const string V3 = "&publishedfileids[0]=";
 
         // Gets mod info for multiple mods
-        public static List<JObject> GetFileDetails(List<int> modIds)
+        public static List<JObject> GetFileDetails(List<double> modIds)
         {
             try 
             {
@@ -33,7 +33,7 @@ namespace FASTER.Models
         }
 
         // Get mod info for single mod
-        public static JObject GetSingleFileDetails(int modId)
+        public static JObject GetSingleFileDetails(double modId)
         {
             try
             {
